Renames data module to constants module
