+++ src/components/task-edit-card.js
@@ -1,4 +1,4 @@
-import {MONTH_NAMES, DAYS, COLORS} from './../data.js';
+import {MONTH_NAMES, DAYS, COLORS} from './../constants.js';
 import {getTime} from './../utils.js';
 import {createRepeatingDaysComponent} from './days.js';
 import {createColorsComponent} from './colors.js';
